+++ docs/devenv.nix
@@ -47,6 +47,7 @@
       # Serve the mkdocs documentation website with live reload
       exec = "mkdocs serve";
       cwd = config.git.root + "/docs";
+      start.enable = "interactive-shell";
     };
   };